requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
openai==0.27.2
python-dotenv==1.0.0
Werkzeug==2.0.1
pyTigerGraph==1.3.3
pandas
flask-jwt-extended==4.4.4
scikit-learn